+++ driver-core/CHANGELOG.rst
@@ -12,6 +12,7 @@ Merged from 2.0 branch:
 - [bug] Release connections on ResultSetFuture#cancel (JAVA-407)
 - [bug] Fix handling of SimpleStatement with values in query builder
   batches (JAVA-393)
+- [bug] Ensure pool is properly closed in onDown (JAVA-417)
 
 
 2.1.0-rc1: